Grandma situation
# Project-1
This project is a loop about a grandson talking to his grandmother. In order for the grandmother to hear the grandson, he has to speak louder. So by speaking louder means by typing in all caps. When you respond in all caps the grandmother will respond about a random year. In order to end the loop you have to tell her bye in all caps.